Change uniroot to uniroot.integer for target_param == "n"
When searching for a sample size, uniroot wastes HUGE amounts of time narrowing down the fractional part of the solution instead of just stopping when it finds the integer part of the solution. uniroot.integer from the ssanv package will only try integer values. In my tests, this leads to about a 10x speedup. If you don't want to rely on ssanv, all uniroot.integer does is a simple binary search, which is easy enough to code.